Add semantic highlighting to REPL.
Depends on https://github.com/idris-lang/Idris-dev/pull/801

This patch does the following: for expression evaluation and :t at the
REPL, it annotates the returned text with colors (as is done in the
console REPL) as well as tooltips. Thus far, these tooltips show
fully-qualified names, but in the future they can show docstrings,
types, etc.