Add menus for interpreter options
Now, settings in the running interpreter can be modified using
menus. Thus far, only showing implicits and error context are supported,
but the system is straightforward to generalize.

This is not done using customize because the typical use case is to
briefly toggle them while debugging, and persistency is not desired.