Add a hack for bad Idris output
Idris is supposed to tell Emacs how much of the file it loaded. In some
cases, it does not. While this has been fixed in Idris [1], older
versions may still be used with idris-mode. This hack works around the
issue and should be removed after a release or two of Idris have passed.

RE #300 - fixes half of it.

[1] https://github.com/idris-lang/Idris-dev/pull/1960