Add a tree info mode
idris-tree-info-mode is used to show trees that result from interaction
with Idris, such as the caller and callee lists. This has been given its
own dedicated buffer, major mode, and command, because trees are
incompatible with the history feature in idris-info-mode.

Additionally, a new face has been added for buffer titles. It has the
same inheritance default as Magit section headers and is used for tree
viewer and metavariable list buffers.

Fixes #324.