Fix display of smaller error regions
When errors were reported over a small span, idris-mode was calulating
the end position relative to the start position instead of relative to
the beginning of the buffer. This fixes that, so that entire buffers
don't get highlighted.

This issue never came up in the past because precise error spans were
rare. 0.9.18 has many more precise spans to support semantic
highlighting, which exposed the issue.