Add the empty list as an explicit package option
This works around an incompatibility between ido-completing-read and
Emacs's own built-in completing-read. The former makes it difficult to
enter the empty string, even though that should always be allowed
according to the docstring for completing-read. The extra option doesn't
seem to get in the way of vanilla completing-read or the Helm version.

Fixes #355 so far as I can tell.