Add support for new Idris command :make-case
This command allows the generation of case blocks in holes. The mode is
set up to use the same keybinding to case-split a pattern variable or a
hole, to avoid needing to learn more bindings. Additionally, the context
menus for holes have been updated.

This requires an up-to-date Idris, but the only regression on an old
Idris is that the case-split command will give a more confusing error
message when run on a hole.