Use annotations for Eldoc instead of compiler
Now, when the compiler annotates source files, the Eldoc strings are
pre-computed. This makes Eldoc lookup instantaneous, and allows it to
work when the compiler is busy doing other things. It also makes the
docstring tooltips of annotated buffers useful in terminal Emacs, by
exposing this information in another way.

Finally, even names that can't be referenced (e.g. the contents of where
blocks) now get eldoc support.

Fixes #251 by not round-tripping to the compiler.

Fixes #279 the same way.