Delete semantic highlights on modification
When a user edits a source file with semantic highlighting, it should
disappear, as that highlighting no longer reflects the truth about that
identifier.

This change also prevents some absurd situations where a great deal of
code was typed in a region that had a semantic highlight, but had
nothing to do with that hightlight. This caused things like
documentation lookup to return absurd results.

Fixes #411.