Jump to last Idris Code buffer when we quit buffer
With current version, when Idris buffer is displayed in window
that already contains other buffer then
on pressing "q" to quit Idris buffer the focus
stays in the window of the Idris buffer.

This change ensures that the focus returns to the buffer
and window from where we have most likely
came to the killed Idris buffer.