Make empty string always return value from `idris-process-filter`
Why:
In commit: 74731b53caa6b2a2a was introduced regression by moving the ""
inside of the if expression causing it being return value only in some
cases.
The return value from idris-process-filter is written into idris-process
buffer and
after commit 74731b5 this become mostly the output of idris-connect
which consequently returns idris-words-of-encouragement.
Because now the *idris-process* buffer is not empty it is moved to the
front and interrupts users workflow.

Downside of this fix is that the words of encouragement
may be lost from users sight due to another messages when loading
file to Idris. This will be addressed in separate commit.