Mark `idris-list-holes-on-load` as obsolete in favour of `idris-list-holes`
Why:
The `idris-list-holes-on-load` was defined as interactive and such
as invokable by user but with nondeterministic behaviour affected
by `idris-hole-show-on-load`.
Removing it in future will simplify the public api of idris-mode
and reduce maintainance costs.

This commit also updates docs for `idris-prover-success-hook`
and `idris-load-file-success-hook` to make explicit that
they are also affected by the `idris-hole-show-on-load` variable.