Check for semantic property `key` having non empty string value
before assigning the value to `idris-name-key` text property.

Why:
Previously the value was wrapped with `{{{{{ VAL }}}}}` even if it was empty but
later we check for presence of the value as it is non empty we use the value
stored in the idris-name-key property.
```
idris-context-menu-items
...
ref (or (plist-get plist `idris-name-key) (plist-get plist `idris-ref)
```
not using the fallback `idris-ref property value.