Hide notes buffer whe no notes available.
This revisits https://github.com/idris-hackers/idris-mode/pull/148
and improves the user experience when Idris does not report any errors.

Previously in such case the user is left with empty buffer in some of the windows.
After this change if there is nothing to display the window
the notes either display previous buffer or it is closed completely if notes
buffer was the only buffer displayed in it.