Suppress Idris errors when loading a file in `idris-switch-to-repl`
This improves the user workflow.

When writing Idris code, it is often useful to switch back and forth
to the REPL (e.g., to test functions, look up docs, or check types),
even when the code is in a broken state.
However, when an error occurs, switching to the REPL is interrupted,
leaving the point in the original buffer and forcing the user
to dismiss the Idris notes.